requirements.txt
construct==2.10.67
glfw==2.4.0
imgui==1.4.0
mercury-engine-data-structures==0.6.0
PyOpenGL==3.1.5
